add serializer for binary xml text + tweak binary deserializer
Added reading the config to determine trimming options.
